convert eloquent code to doctrine

+++ app/Http/Controllers/BookController.php
@@ -2,21 +2,25 @@
 
 namespace App\Http\Controllers;
 
+use App\Entities\Category;
 use App\Repostiories\BookRepository;
+use App\Repostiories\CategoryRepository;
 use Doctrine\ORM\EntityManager;
 use Illuminate\Http\Request;
 use App\Entities\Book;
-use PhpParser\Node\Expr\Cast\Object_;
+use Illuminate\Support\Facades\DB;
+use PHPUnit\Runner\Exception;
 
 class BookController extends Controller
 {
     protected $em;
     protected $bookRepository;
 
-    public function __construct(EntityManager $em, BookRepository $bookRepository)
+    public function __construct(EntityManager $em, BookRepository $bookRepository, CategoryRepository  $categoryRepository)
     {
         $this->em = $em;
         $this->bookRepository = $bookRepository;
+        $this->categoryRepository = $categoryRepository;
 
     }
 
@@ -24,24 +28,51 @@ class BookController extends Controller
     //add new book
     public function create(Request $request)
     {
-        $book = new Book();
+
 
         $isbn = explode("-", $request->get('isbn'));
         $isValid = count($isbn) == 2 && is_numeric($isbn[0]) && is_numeric($isbn[1]) && $isbn[0] == '978' && strlen($isbn[1]) == 10;
         if ($isValid)
         {
             //$book = Book::create($request->all());
+
+
+            $book = new Book();
+
             $book->setIsbn($request->get('isbn'));
             $book->setTitle($request->get('title'));
             $book->setAuthor($request->get('author'));
-            $book->setCategory($request->get('category'));
             $book->setPrice($request->get('price'));
 
+            $category = explode(",",$request->get('category'));
+
+            foreach ($category as $categoryName)
+            {
+
+                $category = $this->categoryRepository->findByName( trim($categoryName));
+
+
+
+                if (!$category){
+
+                    $category = new Category();
+
+                    $category->setName(trim($categoryName));
+
+                    $this->em->persist($category);
+                    $this->em->flush();
+                }
+
+                $book->addCategory($category);
+
+            }
+
             // save information
+
             $this->em->persist($book);
             $this->em->flush();
 
-            return response()->json($book, 201);
+            return response()->json('saved new book with book id: '.$book->getId(), 201);
         }
         return response()->json(["message" => "Invalid ISBN"], 400);
     }
@@ -49,10 +80,19 @@ class BookController extends Controller
     //search book by author name and category
     public function searchByAuthorAndCategory($author, $category)
     {
-        $books = Book::where('author', '=', $author)
-            ->where('category', '=', $category)
-            ->get('isbn');
-        return response()->json($books, 200);
+//        $books = Book::where('author', '=', $author)
+//            ->where('category', '=', $category)
+//            ->get('isbn');
+
+        $books = $this->bookRepository->getBookByAuthorAndCategory($author,$category);
+
+        /** @var Book $book*/
+        foreach($books as $book){
+            $isbn[] =$book->getIsbn();
+        }
+
+
+        return response()->json($isbn, 200);
     }
 
     //search book by author name
@@ -61,17 +101,18 @@ class BookController extends Controller
 //        $books = Book::where('author', '=', $author)
 //            ->get('isbn');
 
-        $books = $this->bookRepository->findByName( $author);
 
-        dump($books);
+        $books = $this->bookRepository->getBookByAuthor($author);
 
-
-        /** @var Book $book */
-
-        foreach($books as $book) {
-            $isbn = $book->getIsbn();
+        /** @var Book $book*/
+        foreach($books as $book){
+            $isbn[] = $book->getIsbn();
         }
-        return response()->json($isbn, 200);
+
+        //dd($book);
+
+
+        return response()->json( $isbn, 200);
     }
 
     //search book by category
@@ -81,12 +122,13 @@ class BookController extends Controller
 //            ->get('isbn');
 //        return response()->json($categories, 200);
 
-        $books = $this->bookRepository->findByCategory( $category);
+        $books = $this->bookRepository->getBookByCategory($category);
 
-        /** @var Book $book */
-        foreach($books as $book) {
-            $isbn = $book->getIsbn();
+        /** @var Book $book*/
+        foreach($books as $book){
+            $isbn[] =$book->getIsbn();
         }
+
         return response()->json($isbn, 200);
     }
 
@@ -94,33 +136,54 @@ class BookController extends Controller
     public function showCategories()
     {
         //get all categories assigned to books
-        $categories = Book::select('category')
-            ->from('books')
-            ->groupBy('category')
-            ->get('category');
+//        $categories = Book::select('category')
+//            ->from('books')
+//            ->groupBy('category')
+//            ->get('category');
+//
+//        //convert categories to array
+//        $decodedCategories = json_decode($categories, true);
+//
+//        /*find all possible categories.
+//        * It is required as one book can have multiple categories assigned as comma separated value.
+//        * This may include duplicates.*/
+//        $allCategories = array();
+//        for ($i = 0; $i < count($decodedCategories); $i++) {
+//
+//            //split comma separated categories assigned to a book
+//            $tempCategories = reset($decodedCategories[$i]);
+//            $categoryList = explode(",", $tempCategories);
+//
+//            for ($j = 0; $j < count($categoryList); $j++) {
+//                array_push($allCategories, array('category' => trim($categoryList[$j])));
+//            }
+//        }
+//
+//        //find unique categories
+//        $uniqCategories = array_unique($allCategories, SORT_REGULAR);
+//
+//        return response()->json($uniqCategories, 200);
 
-        //convert categories to array
-        $decodedCategories = json_decode($categories, true);
 
-        /*find all possible categories.
-        * It is required as one book can have multiple categories assigned as comma separated value.
-        * This may include duplicates.*/
-        $allCategories = array();
-        for ($i = 0; $i < count($decodedCategories); $i++) {
 
-            //split comma separated categories assigned to a book
-            $tempCategories = reset($decodedCategories[$i]);
-            $categoryList = explode(",", $tempCategories);
+       // $books = $this->bookRepository->findAll();
 
-            for ($j = 0; $j < count($categoryList); $j++) {
-                array_push($allCategories, array('category' => trim($categoryList[$j])));
-            }
+
+         //$book =$book->getCategories()->isEmpty();
+
+
+        $categories = $this->categoryRepository->findAllCategory();
+
+        $categoryList = array();
+
+        foreach($categories as $category){
+          $categoryList[] = $category["name"];
         }
 
-        //find unique categories
-        $uniqCategories = array_unique($allCategories, SORT_REGULAR);
 
-        return response()->json($uniqCategories, 200);
+        return response()->json($categoryList, 200);
+
+
 
     }
 
@@ -129,14 +192,25 @@ class BookController extends Controller
     {
         // $book->delete();
 
+
+
         $book = $this->em->find('App\Entities\Book',$id);
 
+        if($book)
+        {
+            $this->em->remove($book);
+            $this->em->flush();
+
+            return response()->json('book deleted successfully', 200);
+
+        }
+        else
+        {
+           throw  new Exception('Book not found');
+        }
 
 
-        $this->em->remove($book);
-        $this->em->flush();
 
-        return response()->json($book, 200);
 
     }
 
@@ -145,8 +219,10 @@ class BookController extends Controller
     {
 
         $isbn = explode("-", $request->get('isbn'));
+
         $isValid = count($isbn) == 2 && is_numeric($isbn[0]) && is_numeric($isbn[1]) && $isbn[0] == '978' && strlen($isbn[1]) == 10;
-        if ($isValid) {
+        if ($isValid)
+        {
 
            // $book->update($request->all());
 
@@ -155,13 +231,33 @@ class BookController extends Controller
             $book->setIsbn($request->get('isbn'));
             $book->setTitle($request->get('title'));
             $book->setAuthor($request->get('author'));
-            $book->setCategory($request->get('category'));
             $book->setPrice($request->get('price'));
 
+
+            $category = explode(",",$request->get('category'));
+
+            $book->resetCategory();
+            foreach ($category as $categoryName)
+            {
+
+                $category = $this->categoryRepository->findByName( trim($categoryName));
+
+                if (!$category){
+
+                    $category->setName(trim($categoryName));
+
+                    $this->em->persist($category);
+                    $this->em->flush();
+                }
+
+                $book->addCategory($category);
+
+            }
+
             // save information
+
             $this->em->persist($book);
             $this->em->flush();
-
 
             return response()->json($book, 200);
         }
